requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
streamlit==1.39.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0
PyPDF2==3.0.1
openai==1.45.0
pydantic==2.8.2
python-multipart==0.0.9
numpy==1.26.4
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.2
